fastapi/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pyjwt==2.3.0
websockets==10.1